% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annot_data}
\alias{annot_data}
\title{Annotation example data}
\format{A \code{\link[GenomicRanges]{GRanges}} object.}
\usage{
annot_data
}
\description{
Data containing gene annotation.
}
\keyword{datasets}

