% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bs_data}
\alias{bs_data}
\title{Bismark Cov formatted BS-Seq example data}
\format{A \code{\link[GenomicRanges]{GRanges}} object.}
\usage{
bs_data
}
\description{
Data coming from BS-Seq experiments in Bismark Cov format.
}
\keyword{datasets}

