% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_encode_caltech_prom_reg.R
\name{create_encode_caltech_prom_reg}
\alias{create_encode_caltech_prom_reg}
\title{Create promoter regions from ENCODE Caltech RNA-Seq data}
\usage{
create_encode_caltech_prom_reg(rna_data, chrom_size = NULL, upstream = -100,
  downstream = 100)
}
\arguments{
\item{rna_data}{data.frame object containing the RNA-Seq data.}

\item{chrom_size}{Optional data.frame containing the chromosome sizes.}

\item{upstream}{Integer defining the length of bp upstream of TSS.}

\item{downstream}{Integer defining the length of bp downstream of TSS.}
}
\value{
The created promoter regions stored in a
 \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
\code{create_encode_caltech_prom_reg} creates promoter regions from ENCODE
 Caltech RNA-Seq data. Using the TSS of RNA-Seq data as ground truth labels
 we create promoter regions \code{N} bp upstream and \code{M} bp
 downstream of TSS.
}
\examples{
# Load the RNA-Seq example dataset
rna_data <- rnaseq_data
promoters <- create_encode_caltech_prom_reg(rna_data)

}
\seealso{
\code{\link{read_rna_encode_caltech}}
}

