% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meth_reg.R
\name{create_meth_reg}
\alias{create_meth_reg}
\title{Create methylation regions for each gene promoter.}
\usage{
create_meth_reg(bs_data, prom_data, upstream = -100, downstream = 100,
  num_CpG = 1, sd_thresh = 0, ignore_strand = FALSE, fmin = -1,
  fmax = 1)
}
\arguments{
\item{bs_data}{\code{\link[GenomicRanges]{GRanges}} object containing the
BS-Seq data.}

\item{prom_data}{\code{\link[GenomicRanges]{GRanges}} object containing
the processed RNA-Seq data, i.e. keeping the promoter regions around TSS.}

\item{upstream}{Integer defining the length of bps upstream of TSS.}

\item{downstream}{Integer defining the length of bps downstream of TSS.}

\item{num_CpG}{Optional integer defining the minimum number of CpGs that
have to be in a methylated region. Regions with less than \code{n} CpGs
are discarded.}

\item{sd_thresh}{Optional numeric defining the minimum standard deviation
of the methylation change in a region.}

\item{ignore_strand}{Logical, whether or not to ignore strand information.}

\item{fmin}{Optional minimum range value for region location scaling.}

\item{fmax}{Optional maximum range value for region location scaling.}
}
\value{
A \code{meth_regions} object containing the following information:
 \itemize{
   \item{ \code{meth_data}: A list containing methylation data, where each
     each entry in the list consists of an L X 3 dimensional matrix, where:
     \enumerate{
       \item{ 1st column: Contains the locations of CpGs relative to TSS,
         where the range (min, max) of possible values is given, by the
         inputs fmin and fmax.
       }
       \item{ 2nd column: Contains the total reads of the CpG in the
         corresponding location.}
       \item{ 3rd column: Contains the methylated reads of the CpG in the
         corresponding location.}
     }
   }
   \item{ \code{prom_ind}: A vector storing the corresponding promoter
     indices.}
 }
}
\description{
\code{create_meth_reg} creates methylation regions using a combination
 of BS-Seq and RNA-Seq data. The BS-Seq data give information for the
 methylated CpGs individually, and the RNA-Seq data are used to locate the
 TSS of each gene and its promoter region.
}
\examples{
# Load the RNA-Seq example dataset
rna_data <- rnaseq_data
prom_data <- create_encode_caltech_prom_reg(rna_data, upstream=-2000, downstream=2000)
bs_data <- rrbs_data

meth_regions <- create_meth_reg(bs_data, prom_data)

}
\seealso{
\code{\link{read_rna_encode_caltech}},
 \code{\link{read_bs_encode_haib}},
 \code{\link{create_encode_caltech_prom_reg}}
}

