% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meth_regions.R
\name{create_meth_regions}
\alias{create_meth_regions}
\title{Create methylation regions for each gene promoter.}
\usage{
create_meth_regions(rrbs_data, promoter_data, upstream, downstream,
  num_CpG = 0, sd_thresh = 0, fmin = -1, fmax = 1)
}
\arguments{
\item{rrbs_data}{\code{\link[GenomicRanges]{GRanges}} object containing the
RRBS data.}

\item{promoter_data}{\code{\link[GenomicRanges]{GRanges}} object containing
the processed RNA-Seq data, i.e. keeping the promoter regions around the TSS.}

\item{upstream}{Integer defining the length of bps upstream of TSS.}

\item{downstream}{Integer defining the length of bps downstream of TSS.}

\item{num_CpG}{Optional integer defining the minimum number of CpGs that
have to be in a methylated region. Regions with less than \code{n} CpGs
are discarded.}

\item{sd_thresh}{Opotional numeric defining the minimum standard deviation
of the methylation change in a region.}

\item{fmin}{Optional minimum range value for region location scaling.}

\item{fmax}{Optional maximum range value for region location scaling.}
}
\value{
A \code{methRegions} object which contains the following information:
 \itemize{
   \item{ \code{meth_data}: A list containing the methylation data, where each
     each entry in the list consists of a 3 X n dimensional matrix, where:
     \enumerate{
       \item{ 1st row: Contains the locations of the CpGs relative to TSS, where
         the range (min, max) of possible values is given, by the inputs fmin and fmax.
       }
       \item{ 2nd row: The total reads of the CpG in the corresponding location.}
       \item{ 3rd row: The methylated reads of the CpG in the corresponding location.}
     }
   }
   \item{ \code{prom_ind}: A vector storing the corresponding promoter indices.}
 }
}
\description{
\code{create_meth_regions} creates methylation regions data using a
combination of RRBS and RNA-Seq data. The RRBS data give information for
the methylated CpGs individually, and the RNA-Seq data are used to locate
the Transcription Start Sites (TSS) of each gene and its promoter region.
The RRBS and RNA-Seq data should be a \code{\link[GenomicRanges]{GRanges}}
object.
}
\seealso{
\code{\link{read.rnaseq}}, \code{\link{read.rrbs}}, \code{\link{create_prom_regions}}
}

