% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prom_regions.R
\name{create_prom_regions}
\alias{create_prom_regions}
\title{Create promoter regions using RNA-Seq data}
\usage{
create_prom_regions(data, chrom_size = NULL, upstream = 100,
  downstream = 100)
}
\arguments{
\item{data}{List object containing the RNA-Seq data.}

\item{chrom_size}{Optional data frame containing the chromosome sizes.}

\item{upstream}{Integer defining the length of bps upstream of TSS.}

\item{downstream}{Integer defining the length of bps downstream of TSS.}
}
\value{
the created promoters stored in a \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
\code{create_prom_regions} creates promoter regions data using Transcription
Start Sites (TSS) of RNA-Seq data as ground truth labels. Using the TSS
data we create promoter regions \code{n} base pairs upstream and downstream
of TSS. The RNA-Seq data should be a \code{\link{list}} object.
}
\examples{
# Load the RNA-Seq example dataset
data <- rnaseq_data
promoters <- create_prom_regions(data)

}
\seealso{
\code{\link{read.rnaseq}}
}

