% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tss_reg.R
\name{create_tss_reg}
\alias{create_tss_reg}
\title{Create TSS regions using RNA-Seq data}
\usage{
create_tss_reg(data, chrom_size = NULL, upstream = -100, downstream = 100)
}
\arguments{
\item{data}{The TSS data}

\item{chrom_size}{Optional data.frame containing the chromosome sizes.}

\item{upstream}{Integer defining the length of bp upstream of TSS.}

\item{downstream}{Integer defining the length of bp downstream of TSS.}
}
\value{
The created promoters stored in a \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
\code{create_tss_reg} creates promoter regions data using Transcription
Start Sites (TSS) of RNA-Seq data as ground truth labels. Using the TSS
data we create promoter regions \code{n} bps upstream and \code{m} bps
downstream of TSS. The RNA-Seq data should be a \code{\link{list}} object.
}

