% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_meth_expr_data.R
\name{diff_meth_expr_data}
\alias{diff_meth_expr_data}
\title{Wrapper method for processing RRBS and RNA-Seq HTS data.}
\usage{
diff_meth_expr_data(rrbs_files, rnaseq_files, chrom_size_file = NULL,
  upstream = -100, downstream = 100, num_CpG = 1, sd_thresh = 0,
  is_del_chrom = FALSE, rrbs_cov = 0, fmin = -1, fmax = 1)
}
\arguments{
\item{rrbs_files}{The names of the RRBS '.bed' formatted file to read
data values from.}

\item{rnaseq_files}{The names of the RNA-Seq '.bed' formatted file to
read data values from.}

\item{chrom_size_file}{Optional name of the file containing genome
chromosome sizes}

\item{upstream}{Integer defining the length of bps upstream of TSS.}

\item{downstream}{Integer defining the length of bps downstream of TSS.}

\item{num_CpG}{Optional integer defining the minimum number of CpGs that
have to be in a methylated region. Regions with less than \code{n} CpGs
are discarded.}

\item{sd_thresh}{Optional numeric defining the minimum standard deviation
of the methylation change in a region.}

\item{is_del_chrom}{Optional logical: if TRUE, delete chromosomes X,
Y and M.}

\item{rrbs_cov}{Optional integer to disacrd low coverage reads.}

\item{fmin}{Optional minimum range value for region location scaling.}

\item{fmax}{Optional maximum range value for region location scaling.}
}
\value{
A \code{methExpr} object which contains, among others the following
 information:
 \itemize{
   \item{ \code{meth_data}: A list containing the methylation data, where each
     each entry in the list consists of an L X 3 dimensional matrix, where:
     \enumerate{
       \item{ 1st column: Contains the locations of the CpGs relative to TSS,
         where the range (min, max) of possible values is given, by the
         inputs fmin and fmax.
       }
       \item{ 2nd column: The total reads of the CpG in the corresponding
         location.}
       \item{ 3rd column: The methylated reads of the CpG in the corresponding
         location.}
     }
   }
   \item{ \code{expr_data}: A \code{\link[GenomicRanges]{GRanges}} object containing the
   corresponding gene expression data for each entry of the \code{meth_data} list.}
 }
}
\description{
\code{diff_meth_expr_data} is a wrapper method for processing HTS data and
returning the methylation promoter regions and the corresponding gene
expression levels for those promoter regions.
}

