% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19_data}
\alias{hg19_data}
\title{Human Genome Assembly 19 chromosome sizes.}
\format{A data frame with 1 variable: \code{size}}
\usage{
hg19_data
}
\description{
Data containing the human genome assembly 19 (hg19) chromosome sizes.
}
\keyword{datasets}

