% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_meth_data.R
\name{meth_data}
\alias{meth_data}
\title{Generate methylation data}
\usage{
meth_data(rrbs_data, promoter_data, upstream, downstream, num_CpG = 0,
  sd_thresh = 0, fmin = -1, fmax = 1)
}
\description{
Generate methylation data
}

