% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_prom_cpg_density}
\alias{plot_prom_cpg_density}
\title{Plot promoter CpG density}
\usage{
plot_prom_cpg_density(obj)
}
\arguments{
\item{obj}{Object of type 'methExpr'}
}
\description{
Plot a density of number of CpGs that are found in each promoter region.
}

