% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_prom_cpg_hist}
\alias{plot_prom_cpg_hist}
\title{Plot promoter CpG histogram}
\usage{
plot_prom_cpg_hist(obj)
}
\arguments{
\item{obj}{Object of type 'methExpr'}
}
\description{
Plot a histogram of number of CpGs that are found in each promoter region.
}

