% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_diff_beatson.R
\name{process_diff_beatson}
\alias{process_diff_beatson}
\title{Process differential BS-Seq and RNA-Seq Beatson HTS data}
\usage{
process_diff_beatson(bs_contr_files, bs_treat_files, rna_contr_files,
  rna_treat_files, chrom_size_file = NULL, chr_discarded = NULL,
  upstream = -100, downstream = 100, min_bs_cov = 0, max_bs_cov = 1000,
  cpg_density = 1, sd_thresh = 0, ignore_strand = FALSE, fmin = -1,
  fmax = 1)
}
\arguments{
\item{bs_contr_files}{The name of the control BS-Seq '.bed' formatted files
to read data values from.}

\item{bs_treat_files}{The name of the treatment BS-Seq '.bed' formatted
files to read data values from.}

\item{rna_contr_files}{The name of the control RNA-Seq '.bed' formatted
files to read data values from.}

\item{rna_treat_files}{The name of the treatment RNA-Seq '.bed' formatted
files to read data values from.}

\item{chrom_size_file}{Optional name of the file containing genome
chromosome sizes}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{upstream}{Integer defining the length of bp upstream of TSS.}

\item{downstream}{Integer defining the length of bp downstream of TSS.}

\item{min_bs_cov}{The minimum number of reads mapping to each CpG site.}

\item{max_bs_cov}{The maximum number of reads mapping to each CpG site.}

\item{cpg_density}{Optional integer defining the minimum number of CpGs that
have to be in a methylated region. Regions with less than \code{n} CpGs are
discarded.}

\item{sd_thresh}{Optional numeric defining the minimum standard deviation of
the methylation change in a region. This is used to filter regions with no
methylation change.}

\item{ignore_strand}{Logical, whether or not to ignore strand information.}

\item{fmin}{Optional minimum range value for region location scaling. Under
this version, this parameter should be left to its default value.}

\item{fmax}{Optional maximum range value for region location scaling. Under
this version, this parameter should be left to its default value.}
}
\value{
A \code{diff_processHTS} object which contains among others the
 following information:
 \itemize{
   \item{ \code{methyl_region}: A list containing the methylation regions,
     where each each entry in the list consists of an L X 3 dimensional
     matrix, where:
     \enumerate{
       \item{ 1st col: Contains the locations of the CpGs relative to TSS,
         where the range (min, max) of possible values is given, by the
         inputs fmin and fmax.
       }
       \item{ 2nd col: The total reads of the CpG in the corresponding
         location.}
       \item{ 3rd col: The methylated reads of the CpG in the corresponding
         location.}
     }
   }
   \item{ \code{prom_region}: A \code{\link[GenomicRanges]{GRanges}} object
     containing corresponding annotated promoter regions for each entry of
     the \code{methyl_region} list..

   }
   \item{ \code{rna_data}: A \code{\link[GenomicRanges]{GRanges}} object
     containing the corresponding RNA-Seq data for each entry of the
     \code{methyl_region} list.}
 }
}
\description{
\code{process_diff_beatson} is a wrapper method ...
}
\examples{
# Get the location of the files
bs_file <- system.file("extdata", "bism_rep1.bed", package = "processHTS")
rna_file <- system.file("extdata", "rna_beatson.bed", package = "processHTS")
#data <- process_beatson(bs_file, rna_file)

}

