% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_haib_caltech.R
\name{process_haib_caltech}
\alias{process_haib_caltech}
\title{Wrapper method for processing ENCODE HAIB and Caltech HTS data.}
\usage{
process_haib_caltech(bs_files, rna_files, chrom_size_file = NULL,
  chr_discarded = NULL, upstream = -100, downstream = 100,
  min_bs_cov = 0, max_bs_cov = 1000, cpg_density = 1, sd_thresh = 0,
  ignore_strand = FALSE, fmin = -1, fmax = 1)
}
\arguments{
\item{bs_files}{The name of the BS-Seq '.bed' formatted file to read
data values from.}

\item{rna_files}{The name of the RNA-Seq '.bed' formatted file to
read data values from.}

\item{chrom_size_file}{Optional name of the file containing genome
chromosome sizes}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{upstream}{Integer defining the length of bp upstream of TSS.}

\item{downstream}{Integer defining the length of bp downstream of TSS.}

\item{min_bs_cov}{The minimum number of reads mapping to each CpG site.}

\item{max_bs_cov}{The maximum number of reads mapping to each CpG site.}

\item{cpg_density}{Optional integer defining the minimum number of CpGs that
have to be in a methylated region. Regions with less than \code{n} CpGs
are discarded.}

\item{sd_thresh}{Optional numeric defining the minimum standard deviation
of the methylation change in a region.}

\item{ignore_strand}{Logical, whether or not to ignore strand information.}

\item{fmin}{Optional minimum range value for region location scaling.}

\item{fmax}{Optional maximum range value for region location scaling.}
}
\value{
A \code{methExpr} object which contains, among others the following
 information:
 \itemize{
   \item{ \code{meth_data}: A list containing the methylation data, where each
     each entry in the list consists of an L X 3 dimensional matrix, where:
     \enumerate{
       \item{ 1st column: Contains the locations of the CpGs relative to TSS,
         where the range (min, max) of possible values is given, by the
         inputs fmin and fmax.
       }
       \item{ 2nd column: The total reads of the CpG in the corresponding
         location.}
       \item{ 3rd column: The methylated reads of the CpG in the corresponding
         location.}
     }
   }
   \item{ \code{expr_data}: A \code{\link[GenomicRanges]{GRanges}} object containing the
   corresponding gene expression data for each entry of the \code{meth_data} list.}
 }
}
\description{
\code{process_haib_caltech} is a wrapper method for processing HTS data and
returning the methylation promoter regions and the corresponding gene
expression levels for those promoter regions.
}
\examples{
# Get the location of the files
rrbs_file <- system.file("extdata", "rrbsH1hESC.bed", package = "processHTS")
rnaseq_file <- system.file("extdata", "rnaseqH1hESC.bed", package = "processHTS")
data <- process_haib_caltech(rrbs_file, rnaseq_file)

}

