% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHTS.R
\docType{package}
\name{processHTS}
\alias{.datatable.aware}
\alias{processHTS}
\alias{processHTS-package}
\title{\code{processHTS}: A package for processing HTS data and files}
\format{An object of class \code{logical} of length 1.}
\usage{
.datatable.aware
}
\description{
The processHTS package provides three categories of important functions.
}
\section{\code{read} functions}{

\code{\link{read_chrom_size}} for reading chromosome sizes file,
\code{\link{read_rna_encode_caltech}} read RNA-Seq bed formatted data file,
\code{\link{read_bs_encode_haib}} read RRBS bed formatted data file.
}
\keyword{datasets}

