% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rnaseq.R
\name{read.rnaseq}
\alias{read.rnaseq}
\title{Read file containing \code{bed} formatted RNA-Seq data}
\usage{
read.rnaseq(file, is_del_chrom = FALSE, is_list = TRUE)
}
\arguments{
\item{file}{The name of the file to read data values from.}

\item{is_del_chrom}{Logical: if TRUE, delete chromosomes X, Y and M.}

\item{is_list}{Logical: if TRUE, a list is returned, a GRanges
object otherwise.}
}
\value{
a list if \code{is_list} is TRUE, otherwise a
 \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
\code{read.rnaseq} reads a file containing promoter annotation data
from RNA-Seq experiments using the \code{\link{scan}} function. The RNA-Seq
file should be in \code{bed} format, e.g. use \code{gtf2bed} tool if your
initial file is in \code{gtf} format.
}
\examples{
# Get the location of the RNA-Seq file
rnaseq_file <- system.file("extdata", "rnaseq.bed", package = "processHTS")
rnaseq_data <- read.rnaseq(file=rnaseq_file)

}
\seealso{
\code{\link{read.chrom_size}}, \code{\link{read.rrbs}}
}

