% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rrbs_data}
\alias{rrbs_data}
\title{RRBS example data}
\format{A \code{\link[GenomicRanges]{GRanges}} object.}
\usage{
rrbs_data
}
\description{
Data coming from Reduced Representation Bisulfite Sequecning experiments.
}
\keyword{datasets}

