% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHTS.R
\docType{package}
\name{processHTS}
\alias{processHTS}
\alias{processHTS-package}
\title{\code{processHTS}: A package for processing HTS data and files}
\description{
The processHTS package provides three categories of important functions.
}
\section{\code{read} functions}{

\code{\link{read.chrom_size}} for reading human reference genome file,
\code{\link{read.rnaseq}} read RNA-Seq bed formatred data file,
\code{\link{read.rrbs}} read RRBS bed formatted data file.
}

