% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rrbs.R
\name{read.rrbs}
\alias{read.rrbs}
\title{Read file containing \code{bed} formatted RRBS data}
\usage{
read.rrbs(file, is_del_chrom = FALSE, rrbs_cov = 0, is_list = TRUE)
}
\arguments{
\item{file}{The name of the file to read data values from.}

\item{is_del_chrom}{Logical: if TRUE, delete chromosomes X, Y and M.}

\item{rrbs_cov}{Optional integer to disacrd low coverage reads.}

\item{is_list}{Logical: if TRUE a list is returned, otherwise
a GRanges object.}
}
\value{
a list if \code{is_list} is TRUE, otherwise a
 \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
\code{read.rrbs} reads a file containing methylation data from RRBS
experiments using the \code{\link{scan}} function. The RRBS file should be
in \code{bed} format.
}
\examples{
# Get the location of the RRBS file
rrbs_file <- system.file("extdata", "rrbs.bed", package = "processHTS")
data <- read.rrbs(file=rrbs_file, is_list=FALSE)

}
\references{
\url{http://genome.ucsc.edu/cgi-bin/hgTables?db=hg19&hgta_group=regulation&hgta_track=wgEncodeHaibMethylRrbs&hgta_doSchema=describe+table+schema}
}
\seealso{
\code{\link{read.chrom_size}}, \code{\link{read.rnaseq}}
}

