% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annot_beatson.R
\name{read_annot_beatson}
\alias{read_annot_beatson}
\title{Read file containing gene annotation data from Beatson}
\usage{
read_annot_beatson(file, chr_discarded = NULL, is_GRanges = TRUE)
}
\arguments{
\item{file}{The name of the file to read data values from.}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{is_GRanges}{Logical: if TRUE a GRanges object is returned, otherwise
a data.frame object is returned.}
}
\value{
a \code{\link[GenomicRanges]{GRanges}} object if \code{is_GRanges}
 is TRUE, otherwise a \code{\link[data.table]{data.table}} object.
}
\description{
\code{read_annot_beatson} reads a file containing gene annotation data from
Beatson, using the \code{\link{scan}} function.
}
\examples{
# Get the location of the annotated Beatson file
annot_beatson_file <- system.file("extdata", "annot_beatson.bed", package = "processHTS")
annot_beatson_data <- read_annot_beatson(file=annot_beatson_file, is_GRanges=TRUE)

}
\seealso{
\code{\link{read_chrom_size}}, \code{\link{read_bs_encode_haib}}
}

